QNX: guest_shm: hardcode physical address of shared memory at 96G
The inter-VM shared memory in QNX is only accessible from CPU and
cannot be the source or target of DMA transactions.

As a workaround, we alias-map that shared memory as DMA-able to another
region in the guest physical address space when launching the
VMs. Accordingly the drivers in guest OSes need to use that new address
which is accessible to both CPUs and devices.

This patch uses a hardcoded address base(96G) for simplicity. Later we
should switch to a more generic approach such as ACPI SSDT for reporting
the alias maps.

Tracked-On: OAM-119803
Signed-off-by: Junjie Mao <junjie.mao@intel.com>
Signed-off-by: Yuanyuan Zhao <yuanyuan.zhao@intel.com>