QNX: W/A for virtio-input over shmem
For virtio-input devices, writing to config area is defined as selecting
config pages. But in virtio over shmem implement, this operation is not
supported. The W/A is to update config area upon config selection in BE,
and delay a short while after cfg write in FE. This delay does not reduce
performance as it only happens during init stage.

Signed-off-by: Wu Zhou <wu.zhou@intel.com>