dma-buf: Support negotiation between exporter and importer
Add flags field to struct dma_buf_attachment as a means of metadata
sharing for DMA-BUF exporter and importer. DMABUF_ATTACH_FLAG_LMEM is
defined, acting as an indicator of buffer residing in local memory of a
discrete GPU.

Add importer_type_id to struct dma_buf. This allows the exporter to
recognize the importer and respond properly accordingly. Currently, the
only added id is for virtio-GPU.

Tracked-On: OAM-129610
Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>