vos: add udm handshake
Add udm handshake between BE and FE. When the VM which UDM run in shutdown
and BE do not have change to notify FE, running FE can know the BE
status change by check handshake mask.

The handshake mask lays in virtio header in share memory BAR. It is
checked every 2 seconds.

Signed-off-by: Yuanyuan Zhao <yuanyuan.zhao@intel.com>